Lots of things

+++ include/javabind/detail/get_static_field.hpp
@@ -15,7 +15,7 @@ inline jint get_static_field(JNIEnv* env, jclass cls, jfieldID id, tag<jint>)
   return env->GetStaticIntField(cls, id);
 }
 
-inline jint get_static_field(JNIEnv* env, jclass cls, jfieldID id, tag<jdouble>)
+inline jdouble get_static_field(JNIEnv* env, jclass cls, jfieldID id, tag<jdouble>)
 {
   return env->GetStaticDoubleField(cls, id);
 }
